Fixes to a number of modules, some temporary.
I was unable to fix the gocomics.com template, this should be looked at.

comics.com changed their URLs and I've modified the "small" template to
pull the correct URL.  The "big" version should also be looked at.

Achewood is problematic in that sometimes it is a relative URL, but
sometimes it is an absolute URL.  Prepending the base variable obviously
causes this to break.  Perhaps a new feature would be to see if it is a
relative URL and only then prepend a value?